Change item pack size + quantity inserts back to whole number

+++ server/graphql/invoice_line/src/mutations/outbound_shipment_line/unallocated_line/update.rs
@@ -18,7 +18,7 @@ use service::{
 #[graphql(name = "UpdateOutboundShipmentUnallocatedLineInput")]
 pub struct UpdateInput {
     pub id: String,
-    pub quantity: f64,
+    pub quantity: u32,
 }
 
 #[derive(Interface)]
