+++ server/repository/migrations/sqlite/2021-09-05T10-00_create_item_table/up.sql
@@ -6,7 +6,7 @@ CREATE TABLE item
     name TEXT NOT NULL,
     code TEXT NOT NULL,
     unit_id TEXT REFERENCES unit(id),
-    default_pack_size REAL NOT NULL,
+    default_pack_size INTEGER NOT NULL,
     type TEXT NOT NULL,
     -- TODO, this is temporary, remove
     legacy_record TEXT NOT NULL
