+++ server/service/src/invoice_line/outbound_shipment_unallocated_line/insert.rs
@@ -17,7 +17,7 @@ pub struct InsertOutboundShipmentUnallocatedLine {
     pub id: String,
     pub invoice_id: String,
     pub item_id: String,
-    pub quantity: f64,
+    pub quantity: u32,
 }
 
 #[derive(Clone, Debug, PartialEq)]
@@ -103,7 +103,7 @@ fn generate(
         id,
         invoice_id,
         pack_size: 1,
-        number_of_packs: quantity,
+        number_of_packs: quantity as f64,
         item_id,
         item_code: item.code,
         item_name: item.name,
@@ -201,7 +201,7 @@ mod test_insert {
                     id: existing_invoice_line.id.clone(),
                     invoice_id: "".to_owned(),
                     item_id: "".to_owned(),
-                    quantity: 0.0
+                    quantity: 0
                 },
             ),
             Err(ServiceError::LineAlreadyExists)
@@ -215,7 +215,7 @@ mod test_insert {
                     id: new_line_id.clone(),
                     invoice_id: "invalid".to_owned(),
                     item_id: "item_a".to_owned(),
-                    quantity: 0.0
+                    quantity: 0
                 },
             ),
             Err(ServiceError::InvoiceDoesNotExist)
@@ -229,7 +229,7 @@ mod test_insert {
                     id: new_line_id.clone(),
                     invoice_id: mock_inbound_shipment_a().id.clone(),
                     item_id: "item_a".to_owned(),
-                    quantity: 0.0
+                    quantity: 0
                 },
             ),
             Err(ServiceError::NotAnOutboundShipment)
@@ -243,7 +243,7 @@ mod test_insert {
                     id: new_line_id.clone(),
                     invoice_id: mock_allocated_invoice().id.clone(),
                     item_id: "item_a".to_owned(),
-                    quantity: 0.0
+                    quantity: 0
                 },
             ),
             Err(ServiceError::CanOnlyAddLinesToNewOutboundShipment)
@@ -257,7 +257,7 @@ mod test_insert {
                     id: new_line_id.clone(),
                     invoice_id: new_outbound_shipment.id.clone(),
                     item_id: "invalid".to_owned(),
-                    quantity: 0.0
+                    quantity: 0
                 },
             ),
             Err(ServiceError::ItemNotFound)
@@ -270,7 +270,7 @@ mod test_insert {
                     id: new_line_id.clone(),
                     invoice_id: new_outbound_shipment.id.clone(),
                     item_id: mock_item_service_item().id.clone(),
-                    quantity: 0.0
+                    quantity: 0
                 },
             ),
             Err(ServiceError::NotAStockItem)
@@ -283,7 +283,7 @@ mod test_insert {
                     id: "new unallocated line id".to_owned(),
                     invoice_id: mock_new_invoice_with_unallocated_line().id.clone(),
                     item_id: existing_invoice_line.item_id.clone(),
-                    quantity: 0.0
+                    quantity: 0
                 },
             ),
             Err(ServiceError::NotThisStoreInvoice)
@@ -297,7 +297,7 @@ mod test_insert {
                     id: new_line_id.clone(),
                     invoice_id: new_outbound_shipment.id.clone(),
                     item_id: existing_invoice_line.item_id.clone(),
-                    quantity: 0.0
+                    quantity: 0
                 },
             ),
             Err(ServiceError::UnallocatedLineForItemAlreadyExistsInInvoice)
@@ -330,7 +330,7 @@ mod test_insert {
                     id: "new_line".to_owned(),
                     invoice_id: invoice_id.clone(),
                     item_id: item.id.clone(),
-                    quantity: 4.0,
+                    quantity: 4,
                 },
             )
             .unwrap();
